recommendersystem/requirements.txt
appdirs==1.4.4
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
distlib==0.3.1
dnspython==1.16.0
filelock==3.0.12
Flask==1.1.2
googlemaps==4.2.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==0.14.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lightfm==1.15
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mlrose==1.3.0
mysql-connector-python==8.0.19
numpy==1.18.1
Pillow==8.0.1
protobuf==3.6.1
pycodestyle==2.5.0
pylint==2.4.4
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.0
requests==2.23.0
scikit-learn==0.23.2
scikit-surprise==1.1.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
threadpoolctl==2.1.0
urllib3==1.25.8
virtualenv==20.4.3
waitress==2.0.0
Werkzeug==1.0.1
wrapt==1.11.2
